Move pkimetal client to its own package
